move configuration out of ember-configuration and into test/helpers
