+++ tests/test-helper.js
@@ -2,5 +2,40 @@ import resolver from './helpers/resolver';
 import {
   setResolver
 } from 'ember-qunit';
+import QUnit from 'qunit';
+import ENV from 'dummy/config/environment';
+import DS from 'ember-data';
 
 setResolver(resolver);
+
+const QUNIT_PARAMS = QUnit.urlParams;
+
+ENV.EXTEND_PROTOTYPES = QUNIT_PARAMS.extendprototypes;
+ENV.ENABLE_OPTIONAL_FEATURES = QUNIT_PARAMS.enableoptionalfeatures;
+
+QUnit.begin(function() {
+  Ember.RSVP.configure('onerror', function(reason) {
+    // only print error messages if they're exceptions;
+    // otherwise, let a future turn of the event loop
+    // handle the error.
+    if (reason && reason instanceof Error) {
+      Ember.Logger.log(reason, reason.stack);
+      throw reason;
+    }
+  });
+
+  var transforms = {
+    'boolean': DS.BooleanTransform.create(),
+    'date': DS.DateTransform.create(),
+    'number': DS.NumberTransform.create(),
+    'string': DS.StringTransform.create()
+  };
+
+  // Prevent all tests involving serialization to require a container
+  DS.JSONSerializer.reopen({
+    transformFor: function(attributeType) {
+      return this._super(attributeType, true) || transforms[attributeType];
+    }
+  });
+
+});
